fix UI bugs and get required data to search Nodes
